Implement the nearest-neighbor approach to choose heaviest station
Instead of using Voronoi Diagrams to determina the usage load of every
station in the network, use a brutish approach that generates a high
number of random points and gets the closest station to each one.
At the end, the more frequently chosen station is the one chosen as the
heaviest one.

This is not a formal approach and should be replaced eventually with
Voronoi.

However, for v2.0.0, this could yield better results if instead of
the points being random, they are chosen accordingly to a city's
demographical distribution.

These changes are reflected in (defn stations-load) inside station.clj,
and inside (defn nearest-neighbor) inside math.clj.

The functions inside network.clj were modified to reflect the new
implementations. Not significant in the final API.

-- Misc --

Add a helping function dist that calls q/dist using only two vector
arguments instead of four numbers.